added PrivatePage.spec.js
Added a test for private routes to check the correct rendering states
when authenticated / not authenticated.

The test also checks if the routing changes to `/login` when a
private page is accessed unauthenticated.

I removed `halogen` in the process, as its a) not compatible with
`enzyme`s `describeWithDom` api, and b) works not in IE8.